Corregifo profiling

+++ profiling/main.cpp
@@ -34,7 +34,7 @@ bool modo_1(unsigned short int &n,unsigned short int &m){
     // cout<<"Modo 1"<<endl;
     for (unsigned short int i = 0; i < n; i++){
         for (unsigned short int j = 0; j < m; j++){
-            arreglo_modo_1[i][j]=j+i*m;
+            arreglo_modo_1[i][j]=(unsigned int)(j+i*m);
             // cout<<arreglo_modo_1[i][j]<<" ";
         };
         // cout<<"\n";
@@ -50,7 +50,7 @@ bool modo_2(unsigned short int &n,unsigned short int &m){
     // cout<<"Modo 2"<<endl;
     for (unsigned short int i = 0; i < n; i++){
         for (unsigned short int j = 0; j < m; j++){
-            arreglo_modo_2[j+i*m]=j+i*m;
+            arreglo_modo_2[(unsigned int)(j+i*m)]=(unsigned int)(j+i*m);
             // cout<<arreglo_modo_2[j+i*m]<<" ";
 
         };
@@ -67,7 +67,7 @@ bool modo_3(unsigned short int &n,unsigned short int &m){
     // cout<<"Modo 3"<<endl;
     for (unsigned short int i = 0; i < n; i++){
         for (unsigned short int j = 0; j < m; j++){
-            arreglo_modo_3[i+j*n]=i+j*n;
+            arreglo_modo_3[(unsigned int)(i+j*n)]=(unsigned int)(i+j*n);
         };
     };
 
@@ -88,7 +88,7 @@ bool modo_4(unsigned short int &n,unsigned short int &m){
 
 
     // cout<<"Modo 4"<<endl;
-    for (unsigned int i = 0; i < n*m; i++){    
+    for (unsigned int i = 0; i < (unsigned int)(n*m); i++){    
         arreglo_modo_4[i]=i;
         // cout<<arreglo_modo_4[i]<<" ";
